fixed torus
