+++ dapp/src/components/common/NavBar/index.jsx
@@ -31,15 +31,14 @@ const NavBar = ({
   handleDisconnect,
   isLoggedIn
 }) => {
-  const dispatch = useDispatch()
-  const isInCommunityPage = location.pathname.includes('/community/') || location.pathname.includes('/fuse-community/')
-  const isInNestedCommunityPage = isInCommunityPage && !location.pathname.includes('justCreated') && location.pathname.split('/').length > 4
   const isInIssuancePage = location.pathname.includes('/issuance')
 
   if (isInIssuancePage) {
     return null
   }
-
+  const dispatch = useDispatch()
+  const isInCommunityPage = location.pathname.includes('/community/') || location.pathname.includes('/fuse-community/')
+  const isInNestedCommunityPage = isInCommunityPage && !location.pathname.includes('justCreated') && location.pathname.split('/').length > 4
   const modifier = isInCommunityPage
   const withLogo = !isInCommunityPage
